Make the AravisGigEController ready to be installed with setup.py
