+++ cmake/template/programs/AravisGigEController/setup.py.in
@@ -1,0 +1,18 @@
+#!/usr/bin/env python3
+
+from setuptools import setup
+from os.path import normpath # see https://github.com/roboticslab-uc3m/asibot-main/commit/320d03f
+
+moduleName = 'AravisGigEController'
+
+setup(name=moduleName,
+      version='0.0.1',
+      description = 'GUI to control AravisGigE cameras through YARP',
+      author='David Estevez',
+      author_email="destevez@ing.uc3m.es",
+      url='https://github.com/roboticslab-uc3m/yarp-devices/',
+      package_dir={'AravisGigEController': normpath('${CMAKE_SOURCE_DIR}/programs/')},
+      package_data={'AravisGigEController': ['templates/*.ui']},
+      install_requires=['begins==0.9','pyside==1.2.2'],
+      entry_points={'gui_scripts': 'AravisGigEController = AravisGigEController.__main__:start'} # See https://pypi.python.org/pypi/begins/0.8#entry-points
+)
