--- cmake/template/setup.py.in
@@ -1,20 +1,0 @@
-#!/usr/bin/env python3
-
-try:
-	from setuptools import setup
-except ImportError:
-	from distutils.core import setup
-from os.path import normpath # see https://github.com/roboticslab-uc3m/asibot-main/commit/320d03f
-from os.path import join
-
-moduleName = 'AravisGigEController'
-
-setup(name = moduleName,
-        version = '0.0.1',
-        description = 'GUI to control AravisGigE cameras through YARP',
-        author = 'David Estevez',
-        url = 'https://github.com/roboticslab-uc3m/yarp-devices/',
-        package_dir = {'': normpath('${CMAKE_SOURCE_DIR}/programs/AravisGigEController/')},
-        scripts = [join(normpath('${CMAKE_SOURCE_DIR}/programs/AravisGigEController/'), moduleName+'.py')],
-        install_requires = ['begins==0.9','pyside>=1.2.2']
-)
