+++ programs/AravisGigEController/templates/AravisGigEControllerGUI.ui
